# Improve Pathname performance